+++ index.html
@@ -234,8 +234,7 @@
       </section>
     </template>
 
-<script src="./js/main.js"></script>
+<script src="./js/main.js" type="module"></script>
+<script src="js/functions.js"></script>
 </body>
 </html>
-
-<script src="js/functions.js"></script>
